% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_notifications.R
\name{check_fdx1_coding}
\alias{check_fdx1_coding}
\title{Check the Level 4 foodex codes (not the descr) in the Consunption data}
\usage{
check_fdx1_coding(data)
}
\arguments{
\item{data}{The uploaded consumption data}
}
\description{
Check the Level 4 foodex codes (not the descr) in the Consunption data
}
