% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_global.R
\name{percent}
\alias{percent}
\title{Turn a vector of numerics into percentages}
\usage{
percent(x, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{...}{Arguments passed to scales::percent()}
}
\value{
A vector of percentages same length as x
}
\description{
Turn a vector of numerics into percentages
}
