% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_global.R
\name{vlookup}
\alias{vlookup}
\title{Excel vlookup}
\usage{
vlookup(this, df, key, value)
}
\arguments{
\item{this}{The value to lookup}

\item{df}{the dataframe to look up}

\item{key}{String length 1? the col name of the df to look up}

\item{value}{String length 1? the col name of the df to return
This comes from the `Goddess`
See https://twitter.com/JennyBryan/status/980978609794895872
https://www.r-bloggers.com/2018/04/an-r-vlookup-not-so-silly-idea/}
}
\description{
Excel vlookup
}
